+++ imio/dms/mail/columns.py
@@ -238,6 +238,34 @@ class ReviewStateColumn(I18nColumn):
         return translate(safe_unicode(state_title), domain=self.i18n_domain, context=self.request)
 
 
+class SessionIdColumn(BaseColumn):
+    """Versions table column to display session id when there are multiple sessions"""
+    weight = 10
+    escape = False
+
+    def renderHeadCell(self):
+        return u'<img src="++resource++imio.esign/parapheo.svg" style="height:1em;vertical-align:middle"> ID'
+
+    def renderCell(self, content):
+        session_id = self.table._session_annotation.get("uids", {}).get(content.UID, None)
+        portal = api.portal.get()
+        dashboard_link = getMultiAdapter((portal, portal.REQUEST), name="parapheo").get_dashboard_link(
+            {"id": session_id}
+        )
+        if session_id is not None and len(self.table._approval.session_ids) > 1:
+            return u'<a href={dashboard_link} title="{title}" class="pdf-session-badge">{session_id}</span>'.format(
+                dashboard_link=dashboard_link,
+                title=translate(
+                    u"Paraphéo session ID: ${session_id}",
+                    domain="imio.dms.mail",
+                    context=content.REQUEST,
+                    mapping={'session_id': session_id},
+                ),
+                session_id=session_id,
+            )
+        return u""
+
+
 class FileActionsColumn(AnnexActionsColumn):
     """IM dashboard. xss ok"""