+++ CHANGES.rst
@@ -54,6 +54,10 @@ Changelog
   [chris-adam]
 - Added actions panel in signing fieldset.
   [chris-adam,sgeulette]
+- Added tree struture on OMVersionsTable for eSigned files.
+  [chris-adam]
+- Added Session ID column on OMVersionsTable.
+  [chris-adam,sgeulette]
 
 3.0 (2021-09-30)
 ----------------